@@ -3,32 +3,33 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "ng serve --port 80",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.1.5",
-    "@angular/common": "^15.2.0",
-    "@angular/compiler": "^15.2.0",
-    "@angular/core": "^15.1.4",
-    "@angular/forms": "^15.1.5",
-    "@angular/material": "^15.1.4",
-    "@angular/platform-browser": "^15.1.5",
-    "@angular/platform-browser-dynamic": "^15.1.4",
-    "@angular/router": "^15.2.0",
-    "flowbite": "^1.6.3",
+    "@angular/animations": "^15.2.1",
+    "@angular/common": "^15.2.1",
+    "@angular/compiler": "^15.2.1",
+    "@angular/core": "^15.2.1",
+    "@angular/forms": "^15.2.1",
+    "@angular/material": "^15.2.1",
+    "@angular/platform-browser": "^15.2.1",
+    "@angular/platform-browser-dynamic": "^15.2.1",
+    "@angular/router": "^15.2.1",
+    "primeicons": "^6.0.1",
+    "primeng": "^15.2.0",
     "rxjs": "~7.5.7",
     "uuid": "^9.0.0",
     "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.1.5",
-    "@angular/cdk": "^15.1.4",
-    "@angular/cli": "^15.2.0",
-    "@angular/compiler-cli": "^15.2.0",
+    "@angular-devkit/build-angular": "^15.2.1",
+    "@angular/cdk": "^15.2.1",
+    "@angular/cli": "^15.2.1",
+    "@angular/compiler-cli": "^15.2.1",
     "@types/jasmine": "~4.3.1",
     "@types/uuid": "^9.0.1",
     "autoprefixer": "^10.4.13",
@@ -39,7 +40,7 @@
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
     "postcss": "^8.4.21",
-    "tailwindcss": "^3.2.6",
+    "tailwindcss": "^3.2.7",
     "tslib": "^2.5.0",
     "typescript": "~4.9.5"
   }
