@@ -4,6 +4,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Http\Request;
 
@@ -17,7 +18,9 @@ class Ranking extends Model
      * @var array<int, string>
      */
     protected $fillable = [
-        'code'
+        'code',
+        'name',
+        'creator'
     ];
 
     /**
@@ -29,34 +32,36 @@ class Ranking extends Model
         'id',
     ];
 
-    public static function assignStudent($id, Request $request)
+    public static function assignStudent(Request $request)
     {
         $data = $request->validate([
-            'code' => 'required|uuid|exists:rankings,code',
-            'points' => 'sometimes|nullable|required|int|gt:0'
+            'student_id' => 'required|exists:students,id',
+            'code' => 'required|uuid|exists:rankings,code'
         ]);
 
         $ranking = Ranking::where('code', $data['code'])->first();
-        $student = Student::find($id);
-
-        if (!($ranking && $student)) {
-            return false;
-        }
+        $student = Student::find($data['student_id']);
 
         return $ranking->students()->attach($student->id, [
-            'points' => $data['points'] ?? 0
+            'points' => 0
         ]);
     }
 
     public static function createFromRequest(Request $request): Ranking
     {
         $data = $request->validate([
-            'code' => 'required|uuid|unique:rankings,code'
+            'code' => 'required|uuid|unique:rankings,code',
+            'name' => 'required|string|unique:rankings,name',
+            'creator' => 'required|exists:teachers,id'
         ]);
 
-        return Ranking::create([
-            'code' => $data['code']
+        $ranking = Ranking::create([
+            'code' => $data['code'],
+            'name' => $data['name'],
+            'creator' => $data['creator']
         ]);
+
+        return Ranking::find($ranking->id);
     }
 
     public static function updateFromRequest($id, Request $request): array|Ranking|null
@@ -65,6 +70,8 @@ class Ranking extends Model
             'ranking_code' => 'required|uuid|unique:rankings,code',
             'student_id' => 'required|int|gt:0',
             'points' => 'sometimes|nullable|required|int|gt:0',
+            'name' => 'required|text|unique:rankings,name',
+            'creator' => 'required|unique:teachers,id'
         ]);
 
         $ranking = Ranking::find($id);
@@ -74,7 +81,11 @@ class Ranking extends Model
         }
 
         $oldRanking = $ranking;
+
         $ranking->id = $data['id'];
+        $ranking->name = $data['name'];
+        $ranking->creator = $data['creator'];
+
         $ranking->pivot->ranking_id = $data['id'];
         $ranking->pivot->student_id = $data['students_id'];
         $ranking->pivot->points = $data['points'] ?? 0;
@@ -88,5 +99,11 @@ class Ranking extends Model
         return $this
             ->belongsToMany(Student::class, 'ranking_student', 'ranking_id', 'student_id')
             ->withPivot('points');
+    }
+
+    public function creator(): BelongsTo
+    {
+        return $this
+            ->belongsTo(Teacher::class);
     }
 }
