@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -16,6 +15,11 @@ return new class extends Migration
         Schema::create('rankings', function (Blueprint $table) {
             $table->id();
             $table->uuid('code')->unique();
+            $table->text('name')->unique();
+
+            $table->unsignedBigInteger('creator');
+            $table->foreign('creator')->on('teachers')->references('id');
+
             $table->timestamps();
         });
     }
