@@ -2,6 +2,7 @@ import {HttpClient, HttpResponse} from '@angular/common/http';
 import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
 import {ICreateRanking} from '../../../models/create/create-ranking';
+import {ICreateStudentAssignation} from '../../../models/create/create-student-assignation';
 import {IRanking} from '../../../models/ranking.model';
 import {IUpdateRanking} from '../../../models/update/update-ranking';
 import {environment} from '../../environments/environment';
@@ -11,12 +12,12 @@ import {CredentialService} from '../credential.service';
   providedIn: 'root',
 })
 export class RankingService {
-  #rankingUrl: string = `${environment.apiUrl}/ranking`;
+  #rankingUrl: string = `${ environment.apiUrl }/ranking`;
   readonly #clientHeaders: { [header: string]: string };
 
   constructor(private http: HttpClient, private credentials: CredentialService) {
     this.#clientHeaders = {
-      'Authorization': `Bearer ${this.credentials.token}`,
+      'Authorization': `Bearer ${ this.credentials.token }`,
       'Accept': 'application/json',
       'Content-Type': 'application/json',
     };
@@ -29,29 +30,53 @@ export class RankingService {
   }
 
   get(id: number): Observable<IRanking | null> {
-    const url: string = `${this.#rankingUrl}/${id}`;
+    const url: string = `${ this.#rankingUrl }/${ id }`;
     return this.http.get<IRanking>(url, {
       headers: this.#clientHeaders,
     });
   }
 
-  create(ranking: ICreateRanking): Observable<HttpResponse<Object>> {
-    return this.http.post(this.#rankingUrl, ranking, {
+  createdBy(id: number): Observable<IRanking[]> {
+    const url: string = `${ this.#rankingUrl }/created_by`;
+    return this.http.post<IRanking[]>(url, { 'id': id }, {
+      headers: this.#clientHeaders,
+    });
+  }
+
+  assignStudent(entity: ICreateStudentAssignation): Observable<HttpResponse<Object>> {
+    const url: string =
+      `${ this.#rankingUrl }/assign`;
+
+    return this.http.post(url, entity, {
       headers: this.#clientHeaders,
       observe: 'response',
     });
   }
 
-  update(id: number, ranking: IUpdateRanking): Observable<HttpResponse<Object>> {
-    const url: string = `${this.#rankingUrl}/${id}`;
-    return this.http.put(url, ranking, {
+  leaderboardsForStudent(id: number): Observable<IRanking[]> {
+    const url: string = `${ this.#rankingUrl }/for/${ id }`;
+    return this.http.get<IRanking[]>(url, {
+      headers: this.#clientHeaders,
+    });
+  }
+
+  create(entity: ICreateRanking): Observable<HttpResponse<Object>> {
+    return this.http.post(this.#rankingUrl, entity, {
+      headers: this.#clientHeaders,
+      observe: 'response',
+    });
+  }
+
+  update(id: number, entity: IUpdateRanking): Observable<HttpResponse<Object>> {
+    const url: string = `${ this.#rankingUrl }/${ id }`;
+    return this.http.put(url, entity, {
       headers: this.#clientHeaders,
       observe: 'response',
     });
   }
 
   delete(id: number): Observable<HttpResponse<Object>> {
-    const url: string = `${this.#rankingUrl}/${id}`;
+    const url: string = `${ this.#rankingUrl }/${ id }`;
     return this.http.delete(url, {
       headers: this.#clientHeaders,
       observe: 'response',
