@@ -6,6 +6,7 @@ import {ITeacher} from '../../../models/teacher.model';
 import {IUpdateTeacher} from '../../../models/update/update-teacher';
 import {environment} from '../../environments/environment';
 import {CredentialService} from '../credential.service';
+import {IUpdatePassword} from '../../../models/update/update-password';
 
 @Injectable({
   providedIn: 'root',
@@ -35,16 +36,23 @@ export class TeacherService {
     });
   }
 
-  create(teacher: ICreateTeacher): Observable<HttpResponse<Object>> {
-    return this.http.post(this.#teacherUrl, teacher, {
+  create(entity: ICreateTeacher): Observable<HttpResponse<Object>> {
+    return this.http.post(this.#teacherUrl, entity, {
       headers: this.#clientHeaders,
       observe: 'response',
     });
   }
 
-  update(id: number, teacher: IUpdateTeacher): Observable<HttpResponse<Object>> {
+  updatePassword(entity: IUpdatePassword) {
+    const url: string = `${this.#teacherUrl}/password`;
+    return this.http.post(url, entity, {
+      headers: this.#clientHeaders,
+    });
+  }
+
+  update(id: number, entity: IUpdateTeacher): Observable<HttpResponse<Object>> {
     const url: string = `${this.#teacherUrl}/${id}`;
-    return this.http.put(url, teacher, {
+    return this.http.put(url, entity, {
       headers: this.#clientHeaders,
       observe: 'response',
     });
