@@ -1,10 +1,10 @@
 export interface ICreateStudent {
-  avatar: string,
-  name: string,
-  surnames: string,
-  nickname: string,
-  email: string,
-  password: string,
-  password_confirmation: string,
-  birth_date: string,
+  avatar: string;
+  name: string;
+  surnames: string;
+  nickname: string;
+  email: string;
+  password: string;
+  password_confirmation: string;
+  birth_date: string;
 }
