@@ -1,10 +1,10 @@
 export interface ICreateTeacher {
-  avatar: string,
-  name: string,
-  surnames: string,
-  nickname: string,
-  email: string,
-  password: string,
-  password_confirmation: string,
-  center: string,
+  avatar: string;
+  name: string;
+  surnames: string;
+  nickname: string;
+  email: string;
+  password: string;
+  password_confirmation: string;
+  center: string;
 }
