@@ -1,6 +1,11 @@
+import {IPivot} from './pivot.model';
+import {IStudent} from './student.model';
+
 export interface IRanking {
-  id: number,
-  code: string,
-  student_id: number,
-  rank: number
+  id: number;
+  code: string;
+  name: string;
+  creator: number;
+  students: IStudent[];
+  pivot: IPivot;
 }
