@@ -1,5 +1,5 @@
 export interface IUpdateRanking {
-  code: string,
-  student_id: number,
-  rank: number
+  code: string;
+  student_id: number;
+  rank: number;
 }
