@@ -1,7 +1,7 @@
 export interface IUpdateStudent {
-  avatar: string,
-  name: string,
-  surnames: string,
-  nickname: string,
-  birth_date: string,
+  avatar: string;
+  name: string;
+  surnames: string;
+  nickname: string;
+  birth_date: string;
 }
