@@ -1,7 +1,7 @@
 export interface IUpdateTeacher {
-  avatar: string,
-  name: string,
-  surnames: string,
-  nickname: string,
-  center: string,
+  avatar: string;
+  name: string;
+  surnames: string;
+  nickname: string;
+  center: string;
 }
